results_fn_bigquery/requirements.txt
functions_framework
google-cloud-bigquery[bqstorage,pandas]
pandas
pandas-gbq